+++ Plugins/UI/Source/NexusUIEditor/Private/NEditorUtilityWidget.cpp
@@ -14,15 +14,48 @@ const FString UNEditorUtilityWidget::WidgetState_TabDisplayText = TEXT("NWidget_
 const FString UNEditorUtilityWidget::WidgetState_TabIconStyle = TEXT("NWidget_TabIconStyle");
 const FString UNEditorUtilityWidget::WidgetState_TabIconName = TEXT("NWidget_TabIconName");
 
+void UNEditorUtilityWidget::PinTemplate(UEditorUtilityWidgetBlueprint* Template)
+{
+	PinnedTemplate = Template;
+	PinnedTemplate->AddToRoot();
+}
+
+void UNEditorUtilityWidget::UnpinTemplate()
+{
+	if (PinnedTemplate != nullptr)
+	{
+		PinnedTemplate->RemoveFromRoot();
+		PinnedTemplate = nullptr;
+	}
+}
+
+bool UNEditorUtilityWidget::IsPersistent() const
+{
+	return bIsPersistent;
+}
+
+FName UNEditorUtilityWidget::GetUserSettingsIdentifier()
+{
+	if (PinnedTemplate != nullptr)
+    {
+    	return PinnedTemplate->GetRegistrationName();
+    }
+    return GetFName();
+}
+
+FString UNEditorUtilityWidget::GetUserSettingsTemplate()
+{
+	if (PinnedTemplate != nullptr)
+	{
+		return PinnedTemplate->GetFullName();
+	}
+	return GetFullName();
+}
+
 void UNEditorUtilityWidget::NativeConstruct()
 {
-	UE_LOG(LogNexusUIEditor, Warning, TEXT("UNEditorUtilityWidget::NativeConstruct: %s"), *GetUserSettingsIdentifier().ToString());
 	Super::NativeConstruct();
 	
-	UE_LOG(LogNexusUIEditor, Warning, TEXT("REGISTER WIDGET"));
-	UNEditorUtilityWidgetSystem::RegisterWidget(this);
-	
-	UE_LOG(LogNexusUIEditor, Warning, TEXT("BIND CLOSE"));
 	// Bind our default behaviour
 	if (OnTabClosedCallback.IsBound())
 	{
@@ -30,27 +63,27 @@ void UNEditorUtilityWidget::NativeConstruct()
 	}
 	OnTabClosedCallback.BindUObject(this, &UNEditorUtilityWidget::OnTabClosed);
 	
-	if (UNEditorUtilityWidgetSystem::IsRebuildingWidgets())
+	if (IsPersistent())
 	{
-		UNEditorUtilityWidgetSystem* System = GEditor->GetEditorSubsystem<UNEditorUtilityWidgetSystem>();
-		System->RestoreTransientWidget(this);
-		UE_LOG(LogNexusUIEditor, Warning, TEXT("REBUILDING WIDGET"));
+		UNEditorUtilityWidgetSystem::RegisterPersistentWidget(this);
+		if (UNEditorUtilityWidgetSystem::IsRebuildingWidgets())
+		{
+			UNEditorUtilityWidgetSystem* System = UNEditorUtilityWidgetSystem::Get();
+			System->RestorePersistentWidget(this);
+		}
 	}
 	
 	// We need to ensure that the window has its icon after all -- this oddly only executes once if you are opening multiple windows at once.
-	UE_LOG(LogNexusUIEditor, Warning, TEXT("MAKE TASK"));
 	UAsyncEditorDelay* DelayedConstructTask = NewObject<UAsyncEditorDelay>();
 	DelayedConstructTask->Complete.AddDynamic(this, &UNEditorUtilityWidget::DelayedConstructTask);
-	UE_LOG(LogNexusUIEditor, Warning, TEXT("START TASK"));
 	DelayedConstructTask->Start(0.01f, 1);
 }
 
 void UNEditorUtilityWidget::NativeDestruct()
 {
-	UE_LOG(LogNexusUIEditor, Warning, TEXT("UNEditorUtilityWidget::NativeDestruct: %s"), *GetUserSettingsIdentifier().ToString());
-	UNEditorUtilityWidgetSystem::UnregisterWidget(this);
+	UNEditorUtilityWidgetSystem::UnregisterPersistentWidget(this);
 	
-	// This stops it from transitioning between maps
+	// This stops it from removing the tab when transitioning between maps
 	if (!IsPersistent())
 	{
 		UEditorUtilitySubsystem* EditorUtilitySubsystem = GEditor->GetEditorSubsystem<UEditorUtilitySubsystem>();
@@ -58,44 +91,42 @@ void UNEditorUtilityWidget::NativeDestruct()
 	}
 	
 	UnpinTemplate();
-
 	Super::NativeDestruct();
+}
+
+FName UNEditorUtilityWidget::GetTabIdentifier()
+{
+	if (PinnedTemplate != nullptr)
+	{
+		return PinnedTemplate->GetRegistrationName();
+	}
+	return GetFName();
 }
 
 // ReSharper disable once CppMemberFunctionMayBeStatic
 // ReSharper disable once CppMemberFunctionMayBeConst
 void UNEditorUtilityWidget::DelayedConstructTask()
 {
-	//PreDelayedConstructTask();
-	UE_LOG(LogNexusUIEditor, Warning, TEXT("UNEditorUtilityWidget::DelayedConstructTask: %s"), *GetUserSettingsIdentifier().ToString());
-	
 	// We need to do this _late_ as the identifier might not be set yet (as it could be based off the pinned template), unless overridden.
 	if (IsPersistent())
 	{
-		UNEditorUtilityWidgetSystem* System = GEditor->GetEditorSubsystem<UNEditorUtilityWidgetSystem>();
+		UNEditorUtilityWidgetSystem* System = UNEditorUtilityWidgetSystem::Get();
 		if (System != nullptr)
 		{
 			if (System->HasWidgetState(GetUserSettingsIdentifier()))
 			{
-				UE_LOG(LogNexusUIEditor, Warning, TEXT("UNEditorUtilityWidget::DelayedConstructTask - HAS STATE"));
 				RestoreWidgetState(this, GetUserSettingsIdentifier(), System->GetWidgetState(GetUserSettingsIdentifier()));
 			}
 			else
 			{
-				UE_LOG(LogNexusUIEditor, Warning, TEXT("UNEditorUtilityWidget::DelayedConstructTask - NEW STATE"));
 				// By this time we have added a proper identifier through restore state
-				System->AddWidgetState(GetUserSettingsIdentifier(), GetUserSettingsTemplate(), GetState());
+				System->AddWidgetState(GetUserSettingsIdentifier(), GetUserSettingsTemplate(), GetWidgetState(this));
 			}
 		}
-	
 	}
-	
-	UE_LOG(LogNexusUIEditor, Warning, TEXT("UNEditorUtilityWidget::DelayedConstructTask (POST): %s"), *GetUserSettingsIdentifier().ToString());
-	
 	
 	FNEditorUtils::UpdateTab(GetTabIdentifier(), GetTabDisplayBrush(), GetTabDisplayText(), OnTabClosedCallback);
 	FNEditorUtils::UpdateWorkspaceItem(GetTabIdentifier(), GetTabDisplayText(), GetTabDisplayIcon());
-	
 	
 	// We need a render to happen so this can be updated
 	UnitScale = GetTickSpaceGeometry().GetAbsoluteSize() / GetTickSpaceGeometry().GetLocalSize();
@@ -103,7 +134,6 @@ void UNEditorUtilityWidget::DelayedConstructTask()
 
 void UNEditorUtilityWidget::OnTabClosed(TSharedRef<SDockTab> Tab)
 {
-	
 	if (IsPersistent() && !IsEngineExitRequested())
 	{
 		GEditor->GetEditorSubsystem<UNEditorUtilityWidgetSystem>()->RemoveWidgetState(GetUserSettingsIdentifier());