+++ Plugins/UI/Source/NexusUIEditor/Public/NEditorUtilityWidgetSystem.h
@@ -6,15 +6,13 @@
 #include "NWidgetState.h"
 #include "NWidgetStateSnapshot.h"
 #include "Engine/DeveloperSettings.h"
+#include "Macros/NEditorSubsystemMacros.h"
 #include "NEditorUtilityWidgetSystem.generated.h"
 
 class UNEditorUtilityWidget;
 
-// TODO Make Get for EditorSubsystem
-
-
 /**
- * System to store state for any Widget that knows of the system.
+ * System to store state for any Widget that knows of the system, and facilitate restoring that state in different scenarios.
  */
 UCLASS(config = EditorPerProjectUserSettings)
 class NEXUSUIEDITOR_API UNEditorUtilityWidgetSystem : public UEditorSubsystem
@@ -23,42 +21,37 @@ class NEXUSUIEDITOR_API UNEditorUtilityWidgetSystem : public UEditorSubsystem
 
 public:
 	GENERATED_BODY()
+	N_EDITOR_SUBSYSTEM(UNEditorUtilityWidgetSystem)
 
 	static bool IsRebuildingWidgets() { return bIsOpeningMap;}
+	static void RegisterPersistentWidget(UNEditorUtilityWidget* Widget);
+	static void UnregisterPersistentWidget(UNEditorUtilityWidget* Widget);
+	static void RestorePersistentWidget(UNEditorUtilityWidget* Widget);
 	
-	static void RegisterWidget(UNEditorUtilityWidget* Widget);
-	static void UnregisterWidget(UNEditorUtilityWidget* Widget);
-	static UNEditorUtilityWidget* GetWidget(const FName& Identifier);
-	static bool HasWidget(const FName& Identifier);
+	static void OnAssetEditorRequestedOpen(UObject* Object);
+	static void OnMapOpened(const FString& Filename, bool bAsTemplate);
 	
 	static UNEditorUtilityWidget* CreateWithState(const FString& InTemplate, const FName& InIdentifier, FNWidgetState& WidgetState);
 
 	virtual void Initialize(FSubsystemCollectionBase& Collection) override;
 	virtual void Deinitialize() override;
 	
-	void OnAssetEditorRequestedOpen(UObject* Object);
-	void OnMapOpened(const FString& Filename, bool bAsTemplate);
-
-	
 	void AddWidgetState(const FName& Identifier, const FString& Template,  const FNWidgetState& WidgetState);
 	void RemoveWidgetState(const FName& Identifier);
 	void UpdateWidgetState(const FName& Identifier,  const FNWidgetState& WidgetState);
 	bool HasWidgetState(const FName& Identifier) const;
 	FNWidgetState& GetWidgetState(const FName& Identifier);
-	
-	void RestoreTransientWidget(UNEditorUtilityWidget* Widget);
-	
+
 protected:	
 	UPROPERTY(config)   
 	FNWidgetStateSnapshot WidgetStates;
 	
-	FNWidgetStateSnapshot TransientStates;
 private:
-	FDelegateHandle OnAssetEditorRequestedOpenHandle;
-	FDelegateHandle OnMapOpenedHandle;
-	
-	static TArray<UNEditorUtilityWidget*> KnownWidgets;
+	static TArray<UNEditorUtilityWidget*> PersistentWidgets;
+	static FNWidgetStateSnapshot PersistentStates;
 	static bool bIsOpeningMap;
 	
+	FDelegateHandle OnAssetEditorRequestedOpenHandle;
+	FDelegateHandle OnMapOpenedHandle;
 };